apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
langchain-anthropic==0.3.0
langgraph==0.2.50
anthropic==0.40.0
prisma==0.15.0
redis==5.0.8
gitpython==3.1.43
pygithub==2.4.0
docker==7.1.0
chromadb==0.5.20
tree-sitter==0.23.0
cryptography==43.0.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.9.0
arq==0.25.0
clerk-backend-api==1.4.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
langgraph-checkpoint-postgres==2.0.13
voyageai==0.3.3
bleach==6.1.0
